change: data type of entities and related methods
Long data change to Integer for ID of entities and related methods.
Swagger present problems to render page when long type and how I do not
expect a lot of records that justify the long was modified.

+++ src/main/java/com/gabo/videoClub/repositories/IBorrowRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IBorrowRepository extends JpaRepository<Borrow, Long> {
+public interface IBorrowRepository extends JpaRepository<Borrow, Integer> {
 }
