+++ src/main/java/com/gabo/videoClub/repositories/IClientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IClientRepository extends JpaRepository<Client, Long> {
+public interface IClientRepository extends JpaRepository<Client, Integer> {
 }
